+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v0.0.12 - in progress
 - Cleanup whitespace in yaml.
-- Resolve duplicated field definitions.
+- Tools to resolve duplicated field definitions.
 
 ## v0.0.11 - 2021-05-18
 - Updated AF and stained microscopy structure schema.